scraper/requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
google-genai
groq
flask==3.0.3
gunicorn==22.0.0
